refactor: move SSR-cache logic to subfolder /rendering-cache
