+++ core-libs/setup/ssr/optimized-engine/rendering-cache/default-cache-entry-size-calculator.ts
