+++ core-libs/setup/ssr/optimized-engine/rendering-cache/index.ts
@@ -1,0 +1,10 @@
+/*
+ * SPDX-FileCopyrightText: 2025 SAP Spartacus team <spartacus-team@sap.com>
+ *
+ * SPDX-License-Identifier: Apache-2.0
+ */
+
+export * from './default-cache-entry-size-calculator';
+export * from './rendering-cache';
+export * from './rendering-cache-size-manager';
+export * from './rendering-cache.model';
